ui on angular

+++ src/main/java/com/example/fmanager/dto/GoalCreateDto.java
@@ -6,7 +6,6 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import jakarta.validation.constraints.Size;
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -23,8 +22,8 @@ public class GoalCreateDto {
     private String name;
     @NotNull(message = "Target amount cannot be null")
     @Positive(message = "Target amount must be positive")
-    private BigDecimal targetAmount;
-    private BigDecimal currentAmount;
+    private double targetAmount;
+    private double currentAmount;
     private LocalDate startDate;
     @NotNull(message = "End date cannot be null")
     @Future(message = "End date must be in the future")
@@ -33,7 +32,7 @@ public class GoalCreateDto {
     private Integer clientId;
 
     public GoalCreateDto(String name,
-                         BigDecimal targetAmount,
+                         double targetAmount,
                          LocalDate startDate,
                          LocalDate endDate,
                          int clientId) {
